requirements.txt
python-dotenv==0.19.2

Flask==2.1.1
gunicorn==20.1.0

python-magic==0.4.24
PyPDF2==1.26.0
pdfplumber==0.5.28
python-docx==0.8.11
openpyxl==3.0.9
requests==2.28.2
charset-normalizer==2.0.7

openai==0.27.0